# Fuzz fix, quote invalid UTF-8 values