--- a/encode_test.go
+++ b/encode_test.go
@@ -53,11 +53,11 @@ func TestEncodeKeyval(t *testing.T) {
 		{key: decimalStringer{5, 9}, value: "v", want: "5.9=v"},
 		{key: (*decimalStringer)(nil), value: "v", err: logfmt.ErrNilKey},
 		{key: marshalerStringer{5, 9}, value: "v", want: "5.9=v"},
-		{key: "k", value: "\xbd", want: "k=\"\\ufffd\""},
-		{key: "k", value: "\ufffd\x00", want: "k=\"\\ufffd\\u0000\""},
-		{key: "k", value: "\ufffd", want: "k=\"\\ufffd\""},
-		{key: "k", value: []byte("\ufffd\x00"), want: "k=\"\\ufffd\\u0000\""},
-		{key: "k", value: []byte("\ufffd"), want: "k=\"\\ufffd\""},
+		{key: "k", value: "\xbd", want: `k="\ufffd"`},
+		{key: "k", value: "\ufffd\x00", want: `k="\ufffd\u0000"`},
+		{key: "k", value: "\ufffd", want: `k="\ufffd"`},
+		{key: "k", value: []byte("\ufffd\x00"), want: `k="\ufffd\u0000"`},
+		{key: "k", value: []byte("\ufffd"), want: `k="\ufffd"`},
 	}
 
 	for _, d := range data {