Summary:
- Fuzz function looked like it was testing the same input twice, I changed that and a reserialization difference was found
- The test failed for the value `"\xbd\x00"` with "reserialized data does not match"; commit message has more details
- Fixed by quoting any invalid UTF-8 string and any string containing the literal `utf8.RuneError`
- Added decode checks on the key making the parser more strict on key input. I'd like some help here, the decode benchmark difference below doesn't seem to justify the edge case it covers
- Probably not ready to merge just yet

New tests results with existing code:

```
--- FAIL: TestDecoder_errors (0.00s)
	decode_test.go:145: got: <nil>, want: logfmt syntax error at pos 5 on line 1: invalid key
	decode_test.go:145: got: <nil>, want: logfmt syntax error at pos 2 on line 1: invalid key
--- FAIL: TestEncodeKeyval (0.00s)
	encode_test.go:71: "k", "\xbd": got 'k=�', want 'k="\ufffd"'
	encode_test.go:71: "k", "�\x00": got 'k="�\u0000"', want 'k="\ufffd\u0000"'
	encode_test.go:71: "k", "�": got 'k=�', want 'k="\ufffd"'
	encode_test.go:71: "k", []byte{0xef, 0xbf, 0xbd, 0x0}: got 'k="�\u0000"', want 'k="\ufffd\u0000"'
	encode_test.go:71: "k", []byte{0xef, 0xbf, 0xbd}: got 'k=�', want 'k="\ufffd"'
--- FAIL: TestMarshalKeyvals (0.00s)
	encode_test.go:135: []interface {}{"�"}: got error: <nil>, want error: invalid key
	encode_test.go:138: []interface {}{"�"}: got '�=null', want ''
	encode_test.go:135: []interface {}{"\xbd"}: got error: <nil>, want error: invalid key
	encode_test.go:138: []interface {}{"\xbd"}: got '�=null', want ''
```

Old benchmarks:

```
BenchmarkDecodeKeyval-8              500           2454245 ns/op         203.73 MB/s
BenchmarkKRDecode-8                  500           2332233 ns/op         214.39 MB/s
BenchmarkEncodeKeyval-8          2000000               704 ns/op              64 B/op          4 allocs/op
```

New benchmarks:

```
BenchmarkDecodeKeyval-8              500           3694369 ns/op         135.34 MB/s
BenchmarkKRDecode-8                 1000           2369236 ns/op         211.04 MB/s
BenchmarkEncodeKeyval-8          2000000               716 ns/op              64 B/op          4 allocs/op
```