@judwhite 2016-11-15
These benchmarks were taken from a different computer.

First PR iteration benchmarks (outdated; for comparison):

```
BenchmarkDecodeKeyval-8              500           3828872 ns/op         130.59 MB/s
BenchmarkKRDecode-8                  500           2522546 ns/op         198.21 MB/s
BenchmarkEncodeKeyval-8          2000000               745 ns/op              64 B/op          4 allocs/op
```

Current `develop` benchmarks:

```
BenchmarkDecodeKeyval-8              500           2598951 ns/op         192.39 MB/s
BenchmarkKRDecode-8                  500           2491137 ns/op         200.71 MB/s
BenchmarkEncodeKeyval-8          2000000               735 ns/op              64 B/op          4 allocs/op
```

Feedback applied benchmarks:

```
BenchmarkDecodeKeyval-8              500           2802288 ns/op         178.43 MB/s
BenchmarkKRDecode-8                  500           2506353 ns/op         199.49 MB/s
BenchmarkEncodeKeyval-8          2000000               734 ns/op              64 B/op          4 allocs/op
```